requirements.txt
opencv-python
numpy
pillow
flask